feat: link edit detail to tmp edit prop page
