--- src/app/(have-nav)/edit-profiles/page.tsx
@@ -1,30 +1,0 @@
-"use client";
-import getCurrentUser from "@/services/getCurrentUser";
-import { useState, useEffect } from "react";
-import Sidebar from "@/components/edit-profile/Sidebar";
-import PersonalPage from "@/components/edit-profile/PersonalPage";
-import OwnerPage from "@/components/edit-profile/OwnerPage";
-import FinancialPage from "@/components/edit-profile/FinancialPage";
-type Tab = "personal" | "financial" | "owner";
-const EditProfile = () => {
-  const [tab, setTab] = useState<Tab>("personal");
-  const switchToPersonal = () => setTab("personal");
-  const switchToFinancial = () => setTab("financial");
-  const switchToOwner = () => setTab("owner");
-
-  return (
-    <div className="flex h-full flex-row ">
-      <Sidebar
-        personal={switchToPersonal}
-        financial={switchToFinancial}
-        owner={switchToOwner}
-      />
-      <div className="m-5">
-        {tab === "personal" && <PersonalPage />}
-        {tab === "financial" && <FinancialPage />}
-        {tab === "owner" && <OwnerPage />}
-      </div>
-    </div>
-  );
-};
-export default EditProfile;
